backend/pyproject.toml
[tool.poetry]
name = "repertoire"
version = "0.1.0"
description = "A release-oriented music server."
authors = ["azuline <azuline@riseup.net>"]
license = "AGPL-3.0"
packages = [
	{ include = "src" },
]

[tool.poetry.scripts]
repertoire = "src.__main__:run"

[tool.poetry.dependencies]
python = "^3.8"
ariadne = "^0.12.0"
click = "^7.1"
huey = "^2.3"
pillow = "^7.2.0"
python-dotenv = "^0.14.0"
quart = "^0.13.1"
requests = "^2.24"
tagfiles = "^0.3.2"
unidecode = "^1.1.1"
voluptuous = "^0.11.7"
yoyo-migrations = "^7.2"

[tool.poetry.dev-dependencies]
autodocsumm = "0.1.13"
black = "=20.8b1"
flake8 = "^3.8"
isort = "^5.4"
pytest = "^6.1.1"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.10.1"
quart-cors = "^0.3.0"
snapshottest = "^0.6.0"
sphinx = "^3.2.1"
sphinx-autodoc-typehints = "^1.11.1"
sphinx-rtd-theme = "^0.5.0"
"sphinxcontrib.httpdomain" = "^1.7.0"

[tool.black]
line-length = 88
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
	| snapshots
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-src = "poetry.masonry.api"
